Move SiblingSubgraph (keep exported same in hugr::views)
